# Performance improvements for acts_like? method